requirements.txt
cryptography==42.0.5
pycryptodome==pycryptodome-3.20.0.dist-info